Bread for Sandwiches

Generic function for extracting an estimator for the bread of sandwiches.

bread(x, ...)

Arguments

x
a fitted model object.

...
arguments passed to methods.

Value

A matrix containing an estimator for the expectation of the negative derivative of the estimating functions, usually the Hessian. Typically, this should be an \(k \times k\) matrix corresponding to \(k\) parameters. The rows and columns should be named as in coef or terms, respectively.

The default method tries to extract vcov and nobs and simply computes their product.

See also

References

Zeileis A (2006), Object-Oriented Computation of Sandwich Estimators. Journal of Statistical Software, 16(9), 1--16. URL http://www.jstatsoft.org/v16/i09/.

Examples

## linear regression
x <- sin(1:10)
y <- rnorm(10)
fm <- lm(y ~ x)

## bread: n * (x'x)^{-1}
bread(fm)

#>             (Intercept)          x
#> (Intercept)   1.0414689 -0.2938577
#> x            -0.2938577  2.0823419

solve(crossprod(cbind(1, x))) * 10

#>                       x
#>    1.0414689 -0.2938577
#> x -0.2938577  2.0823419