Extract Empirical Estimating Functions

Generic function for extracting the empirical estimating functions of a fitted model.

estfun(x, ...)

Arguments

x
a fitted model object.

...
arguments passed to methods.

Value

A matrix containing the empirical estimating functions. Typically, this should be an \(n \times k\) matrix corresponding to \(n\) observations and \(k\) parameters. The columns should be named as in coef or terms, respectively.

The estimating function (or score function) for a model is the derivative of the objective function with respect to the parameter vector. The empirical estimating functions is the evaluation of the estimating function at the observed data (\(n\) observations) and the estimated parameters (of dimension \(k\)).

See also

References

Zeileis A (2006). “Object-Oriented Computation of Sandwich Estimators.” Journal of Statistical Software, 16(9), 1--16. doi: 10.18637/jss.v016.i09

Zeileis A, Köll S, Graham N (2020). “Various Versatile Variances: An Object-Oriented Implementation of Clustered Covariances in R.” Journal of Statistical Software, 95(1), 1--36. doi: 10.18637/jss.v095.i01

Examples

## linear regression
x <- sin(1:10)
y <- rnorm(10)
fm <- lm(y ~ x)

## estimating function: (y - x'beta) * x
estfun(fm)

#>    (Intercept)          x
#> 1   -0.7015578 -0.5903406
#> 2    0.4494980  0.4087274
#> 3    2.2437203  0.3166338
#> 4   -1.0336182  0.7822448
#> 5    1.2040418 -1.1545849
#> 6   -1.4882323  0.4158352
#> 7   -0.5838511 -0.3835824
#> 8   -0.2694160 -0.2665489
#> 9    0.5953328  0.2453477
#> 10  -0.4159175  0.2262679

residuals(fm) * cbind(1, x)

#>                           x
#>  [1,] -0.7015578 -0.5903406
#>  [2,]  0.4494980  0.4087274
#>  [3,]  2.2437203  0.3166338
#>  [4,] -1.0336182  0.7822448
#>  [5,]  1.2040418 -1.1545849
#>  [6,] -1.4882323  0.4158352
#>  [7,] -0.5838511 -0.3835824
#>  [8,] -0.2694160 -0.2665489
#>  [9,]  0.5953328  0.2453477
#> [10,] -0.4159175  0.2262679