A Simple Meat Matrix Estimator

Estimating the variance of the estimating functions of a regression model by cross products of the empirical estimating functions.

meat(x, adjust = FALSE, ...)

Arguments

x
a fitted model object.

adjust
logical. Should a finite sample adjustment be made? This amounts to multiplication with \(n/(n-k)\) where \(n\) is the number of observations and \(k\) the number of estimated parameters.

...
arguments passed to the estfun function.

Details

For some theoretical background along with implementation details see Zeileis (2006).

Value

A \(k \times k\) matrix corresponding containing the scaled cross products of the empirical estimating functions.

See also

References

Zeileis A (2006), Object-Oriented Computation of Sandwich Estimators. Journal of Statistical Software, 16(9), 1--16. doi: 10.18637/jss.v016.i09

Berger S, Graham N, Zeileis A (2017). “Various Versatile Variances: An Object-Oriented Implementation of Clustered Covariances in R.” Journal of Statistical Software, 95(1), 1--36. doi: 10.18637/jss.v095.i01

Examples

x <- sin(1:10)
y <- rnorm(10)
fm <- lm(y ~ x)

meat(fm)

#>             (Intercept)         x
#> (Intercept)   0.5500547 0.3489064
#> x             0.3489064 0.3073370

meatHC(fm, type = "HC")

#>             (Intercept)         x
#> (Intercept)   0.5500547 0.3489064
#> x             0.3489064 0.3073370

meatHAC(fm)

#>             (Intercept)         x
#> (Intercept)   0.5870363 0.4398522
#> x             0.4398522 0.4492865