Isotonic Autocorrelation Function

Autocorrelation function (forced to be decreasing by isotonic regression).

isoacf(x, lagmax = NULL, weave1 = FALSE)

Arguments

x

numeric vector.

lagmax

numeric. The maximal lag of the autocorrelations.

weave1

logical. If set to TRUE isoacf uses the acf.R and pava.blocks function from the original weave package, otherwise R's own acf and isoreg functions are used.

Details

isoacf computes the autocorrelation function (ACF) of x enforcing the ACF to be decreasing by isotonic regression. See also Robertson et al. (1988).

Value

isoacf returns a numeric vector containing the ACF.

References

Lumley T & Heagerty P (1999). “Weighted Empirical Adaptive Variance Estimators for Correlated Data Regression.” Journal of the Royal Statistical Society B, 61, 459--477.

Robertson T, Wright FT, Dykstra RL (1988). Order Restricted Statistical Inference. John Wiley and Sons, New York.

See also

Examples

x <- filter(rnorm(100), 0.9, "recursive")
isoacf(x)
#>   [1]  1.00000000  0.95268062  0.90115399  0.84708223  0.80126997  0.74822136
#>   [7]  0.70196451  0.65208504  0.58910977  0.53329545  0.49152552  0.45477154
#>  [13]  0.41968582  0.38751425  0.34904343  0.31131372  0.28376265  0.26286007
#>  [19]  0.23681644  0.21049727  0.18746369  0.16207847  0.12868771  0.10012758
#>  [25]  0.06241766  0.01658729 -0.02902641 -0.07154847 -0.10347835 -0.13171104
#>  [31] -0.15653912 -0.15653912 -0.15653912 -0.15653912 -0.15653912 -0.15653912
#>  [37] -0.15653912 -0.15653912 -0.15653912 -0.15653912 -0.15653912 -0.15653912
#>  [43] -0.15653912 -0.15653912 -0.15653912 -0.15653912 -0.15653912 -0.15653912
#>  [49] -0.15653912 -0.15653912 -0.15653912 -0.15653912 -0.15653912 -0.15653912
#>  [55] -0.15653912 -0.15653912 -0.15653912 -0.15653912 -0.15653912 -0.15653912
#>  [61] -0.15653912 -0.15653912 -0.15653912 -0.15653912 -0.15653912 -0.15653912
#>  [67] -0.15653912 -0.15653912 -0.15653912 -0.15653912 -0.15653912 -0.15653912
#>  [73] -0.15653912 -0.15653912 -0.15653912 -0.15653912 -0.15653912 -0.15653912
#>  [79] -0.15653912 -0.15653912 -0.15653912 -0.15653912 -0.15653912 -0.15653912
#>  [85] -0.15653912 -0.15653912 -0.15653912 -0.15653912 -0.15653912 -0.15653912
#>  [91] -0.15653912 -0.15653912 -0.15653912 -0.15653912 -0.15653912 -0.15653912
#>  [97] -0.15653912 -0.15653912 -0.15653912 -0.15653912
acf(x, plot = FALSE)$acf
#> , , 1
#> 
#>            [,1]
#>  [1,] 1.0000000
#>  [2,] 0.9526806
#>  [3,] 0.9011540
#>  [4,] 0.8470822
#>  [5,] 0.8012700
#>  [6,] 0.7482214
#>  [7,] 0.7019645
#>  [8,] 0.6520850
#>  [9,] 0.5891098
#> [10,] 0.5332954
#> [11,] 0.4915255
#> [12,] 0.4547715
#> [13,] 0.4196858
#> [14,] 0.3875143
#> [15,] 0.3490434
#> [16,] 0.3113137
#> [17,] 0.2837626
#> [18,] 0.2628601
#> [19,] 0.2368164
#> [20,] 0.2104973
#> [21,] 0.1874637
#>