meat {sandwich}	R Documentation

A Simple Meat Matrix Estimator

Description

Estimating the variance of the estimating functions of a regression model by cross products of the empirical estimating functions.

Usage

meat(x, adjust = FALSE, ...)

Arguments

x
a fitted model object.

adjust
logical. Should a finite sample adjustment be made? This amounts to multiplication with

n/(n-k)

where

n

is the number of observations and

k

the number of estimated parameters.

...
arguments passed to the estfun function.

Details

For some theoretical background along with implementation details see Zeileis (2006).

Value

A

k x k

matrix corresponding containing the scaled cross products of the empirical estimating functions.

References

Zeileis A (2006), Object-Oriented Computation of Sandwich Estimators. Journal of Statistical Software, 16(9), 1–16. URL http://www.jstatsoft.org/v16/i09/.

See Also

sandwich, bread, estfun

Examples

x <- sin(1:10)
y <- rnorm(10)
fm <- lm(y ~ x)

meat(fm)
meatHC(fm, type = "HC")
meatHAC(fm)

[Package sandwich version 2.3-4 Index]